Fix #144: error when binding negative integers to prepared stmt (#146)
The issue here is that we use unsigned int buffers underneath for
binding the prepared statement parameters, and we were doing a simple
assignment like: `buf[1] = x`. This kind of basic `setindex!` calls
convert, which calls something like `UInt64(-1)` and will throw
InexactErrors. But we don't want to conver the value to anything, we
just want the plain bits stored in our buffer, so the solution here is
pretty simple: cast the bits to the right type.