Convert non-binary columns to String type (#153)
* Convert non-binary columns to String type

Fixes #150.

The issue here is that the C API, by default, says that TEXT columns are
of type MYSQL_TYPE_BLOB, from what I gathered because C doesn't really
have a true "string" type, so it doesn't really matter if it's blob or
string anyway. But, upon further investigation, I found in the
MYSQL_FIELD flags field documentation a bit switch to tell whether a
column is in fact binary or not. Using this, we can now detect
accurately whether a column can be treated as a string or should be
returned as a byte vector. The tests added test the return value of
columns defined as TEXT and BLOB and that they are accurately
distinguished on the Julia side.

* fix failing test