Fix calling stored procedures that return resultsets (#154)
Fixes #149.

The issue here is that calling stored procedures that return resultsets
return the resultset, and then an extra, empty resultset (who knows why,
but whatever). In general, we weren't handling when queries returned
multiple resultsets anyway, so this was a more general problem. We now
will return the first resultset as the official resultset and clear out
any additional resultsets if they exist. We should definitely come up
with a better api if people actually need/want multiple resultsets, but
that's a problem for another day.

It turns out we also had problems returning resultsets from prepared
statements, which is the reason for the extra tests and code changes in
prepare.jl.