Implement new `DBInterface.executemultiple` interface function
This allows returning multiple resultsets from a single query call; this
requires setting `multi_statements` to `true` by default, but I couldn't
think of any downside to doing this (vs. the alternative in which a user
finds the function, tries it and immediately fails and has to restart
their connection by setting the flag). Prepared statements don't support
multiple resultsets, and an error is thrown appropriately (from mysql
itself). This PR requires
https://github.com/JuliaDatabases/DBInterface.jl/pull/26 to be merged
and tagged first.