Check whether column is numeric along with unsigned check
Attempt to fix #159. It seems there are cases when `API.isunsigned` will
return true for a timestamp column; this is probably due to some kind of
database or column setting when the table was created, but nonetheless,
it causes issues because if `API.isunsigned`, then we try to call
`unsigned(T)`, which isn't defined for `DateTime`. As opposed to
pirating that definition in MySQL.jl, here we attempt to also check the
`NUM_FLAG`, which should tell us whether the field is numeric or not and
hopefully avoid ever seeing `API.isunsigned` for timestamp columns.