Ignore any host url protocol that includes "mysql://" (#179)
Fixes #170. As far as I can tell, certain projects suggest/require using
"mysql://", "mysqlx://", "jdbc:mysql://", etc. as the protocol portion
of the host url. This ends up with really weird behavior in MySQL.jl
because it assumes there's a plugin that needs to be loaded. So the fix
here is just a simple check if "mysql://" is in the host url and if so,
only use the rest of the string after it.